requirements.txt
aiogram==3.12.0
asyncpg==0.29.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9